ai-agent-sell/handler/requirements.txt
langchain-core
langchain
pandas
pydantic
python-dotenv
langgraph
langchain_openai
rapidfuzz
elasticsearch
